[ci] Add Mac OSX test to CI and refactor android's NDK/SDK install
To achieve this goal we refactor the command to install android's NDK/SDK into the makefile  `ci/android.mk`

Note: in order to not increase the overall build time, we change the python3 tests for `armeabi-v7a`...so we test that architecture with OSX instead of linux.
We also rename the `Python 2` test to `Python 2 armeabi-v7a (with numpy)` to reflect the build arch and numpy's build.